requirements.txt
flask
wikipedia
modal
openai
python-dotenv
requests